Split SongItem component
